secrecy/zeroize: Cleanups for `bytes` feature; remove `Bytes` impls
- Removes the redundant `bytes` feature; imports it as `bytes`
- Removes the `Bytes` impl for `Zeroize`
- Removes the `SecretBytes` type

I will open another PR with a proper solution for `Bytes`. It's a
particularly tricky problem, but one I think has an effective and safe
solution. This PR is just to clear the dust in advance.
