+++ secrecy/Cargo.toml
@@ -23,9 +23,8 @@ travis-ci = { repository = "iqlusioninc/crates", branch = "develop" }
 [dependencies]
 serde = { version = "1", optional = true }
 zeroize = { version = "0.10", path = "../zeroize", default-features = false }
-bytes_crate = { package = "bytes", version = "0.4.12", optional = true }
+bytes = { version = "0.4", optional = true }
 
 [features]
 default = ["alloc"]
 alloc = ["zeroize/alloc"]
-bytes = ["zeroize/bytes", "bytes_crate"]
