+++ secrecy/src/bytes.rs
@@ -1,14 +1,10 @@
-//! Secret bytes
+//! Optional `Secret` wrapper type for the `bytes::BytesMut` crate.
 
 use super::{CloneableSecret, DebugSecret, Secret};
-use bytes_crate::{Bytes, BytesMut};
+use bytes::BytesMut;
 
-/// Secret bytes
-pub type SecretBytes = Secret<Bytes>;
-/// Secret bytes_mut
+/// Alias for `Secret<BytesMut>`
 pub type SecretBytesMut = Secret<BytesMut>;
 
-impl DebugSecret for Bytes {}
-impl CloneableSecret for Bytes {}
 impl DebugSecret for BytesMut {}
 impl CloneableSecret for BytesMut {}
