+++ zeroize/Cargo.toml
@@ -26,9 +26,8 @@ maintenance = { status = "actively-developed" }
 
 [dependencies]
 zeroize_derive = { version = "0.10", path = "../zeroize_derive", optional = true }
-bytes_crate = { package = "bytes", version = "0.4.12", optional = true }
+bytes = { version = "0.4", optional = true }
 
 [features]
 default = ["alloc"]
 alloc = []
-bytes = ["bytes_crate"]
